refactor: 액션 타입, 액션, 리듀서로 분리
