+++ src/pages/ProductDetailPage/index.jsx
@@ -6,7 +6,8 @@ import styled from 'styled-components';
 import { CardDetailButton } from 'components/common/Button';
 import Flex from 'components/common/Flex';
 import LoadingSpinner from 'components/common/LoadingSpinner';
-import { startProduct, setProduct, resetProduct } from 'modules/product';
+
+import { startProduct, setProduct, resetProduct } from 'store/product/actions';
 import { loadProduct } from 'api';
 
 const ProductDetail = () => {
