+++ src/pages/ProductListPage/index.jsx
@@ -4,7 +4,8 @@ import styled from 'styled-components';
 
 import Product from 'components/Product';
 import LoadingSpinner from 'components/common/LoadingSpinner';
-import { setProductList, startProductList } from 'modules/productList';
+
+import { setProductList, startProductList } from 'store/productList/actions';
 import { loadProductList } from 'api';
 
 const ProductList = () => {
