+++ src/store/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import productReducer from './product';
-import productListReducer from 'modules/productList';
+import productReducer from 'store/product/reducer';
+import productListReducer from 'store/productList/reducer';
 
 const rootReducer = combineReducers({ productReducer, productListReducer });
