+++ src/store/product/actionTypes.js
@@ -1,0 +1,5 @@
+export const START_PRODUCT = 'product/START_PRODUCT';
+
+export const SET_PRODUCT = 'product/SET_PRODUCT';
+
+export const RESET_PRODUCT = 'product/RESET_PRODUCT';
