+++ src/store/productList/actionTypes.js
@@ -1,0 +1,3 @@
+export const START_PRODUCT_LIST = 'productList/START_PRODUCT_LIST';
+
+export const SET_PRODUCT_LIST = 'productList/SET_PRODUCT_LIST';
